step9: add s3bucket resources
After we deployed the controller resources to the cluster, now we can create a new s3bucket resource using: `kubectl apply -f resources/`, and the controller will watch this resource and update the s3 bucket public access to the desired state represented by the crd.

Note: the s3 bucket should be created in aws before creating the CRD that represents the bucket.